# apps/function_app/openapi/index.yaml
swagger: '2.0'
info:
  version: 1.10.3
  title: IO API for service Messages
  contact:
    name: PagoPA
    url: https://forum.italia.it/c/progetto-io
  x-logo:
    url: https://io.italia.it/assets/img/io-logo-blue.svg
  description: |
    # Warning Experimental API These apis are experimental and for internal use.
host: api.cd.italia.it
basePath: /api/v1
schemes:
  - https
security:
  - ApiKeyAuth: []
paths:
  /ready:
    get:
      operationId: info
      summary: Info
      responses:
        '200':
          description: Success
        '500':
          description: There was an error in getting info.
          schema:
            $ref: '#/definitions/ProblemJson'
  /resource:
    post:
      operationId: postResource
      summary: Create Resource
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/CreateResource'
          required: true
      responses:
        '200':
          description: Success.
          schema:
            $ref: '#/definitions/CreatedResource'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ProblemJson'
        '401':
          description: Token null or invalid.
        '403':
          description: Request not authorized.
        '404':
          description: Resource not found.
        '500':
          description: There was an error in creating resource.
          schema:
            $ref: '#/definitions/ProblemJson'
      consumes:
        - application/json
      produces:
        - application/json
  /resources/{fiscal_code}/{resource_id}:
    get:
      operationId: getResource
      summary: Get Resource by its Id
      parameters:
        - $ref: '#/parameters/FiscalCodeParam'
        - name: resource_id
          in: path
          type: string
          required: true
          description: The ID of the resource.
      responses:
        '204':
          description: Success.
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/ProblemJson'
        '401':
          description: Token null or invalid.
        '403':
          description: Request not authorized.
        '404':
          description: Resource not found.
        '500':
          description: There was an error in forwarding the notification.
          schema:
            $ref: '#/definitions/ProblemJson'
      consumes:
        - application/json
      produces:
        - application/json
definitions:
  FiscalCode:
    type: string
    description: User's fiscal code.
    format: FiscalCode
    x-import: '@pagopa/ts-commons/lib/strings'
    example: SPNDNL80R13C555X
  CreateResource:
    type: object
    properties:
      fiscal_code:
        $ref: '#/definitions/FiscalCode'
    required:
      - fiscal_code
  CreatedResource:
    type: object
    properties:
      fiscal_code:
        $ref: '#/definitions/FiscalCode'
      resource_id:
        type: string
        minLenght: 1
    required:
      - fiscal_code
      - resource_id
  ProblemJson:
    type: object
    properties:
      type:
        type: string
        format: uri
        description: |-
          An absolute URI that identifies the problem type. When dereferenced,
          it SHOULD provide human-readable documentation for the problem type
          (e.g., using HTML).
        default: about:blank
        example: https://example.com/problem/constraint-violation
      title:
        type: string
        description: |-
          A short, summary of the problem type. Written in english and readable
          for engineers (usually not suited for non technical stakeholders and
          not localized); example: Service Unavailable
      status:
        type: integer
        format: int32
        description: >-
          The HTTP status code generated by the origin server for this occurrence of the problem.
        minimum: 100
        maximum: 600
        exclusiveMaximum: true
        example: 200
      detail:
        type: string
        description: |-
          A human readable explanation specific to this occurrence of the
          problem.
        example: There was an error processing the request
      instance:
        type: string
        format: uri
        description: >-
          An absolute URI that identifies the specific occurrence of the problem.

          It may or may not yield further information if dereferenced.
parameters:
  FiscalCodeParam:
    name: fiscal_code
    in: path
    required: true
    x-example: SPNDNL80R13C555X
    type: string
    description: User's fiscal code.
    format: FiscalCode
    x-import: '@pagopa/ts-commons/lib/strings'
    example: SPNDNL80R13C555X
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-Functions-Key